+++ packages/engine/src/index.ts
@@ -3,6 +3,7 @@ export * from "./renderBuffer";
 export * from "./models/types";
 export * from "./models/bounds";
 export * from "./models/transform";
+export * from "./models/placement";
 export * from "./models/matrix";
 export * from "./models/line";
 export * from "./models/arches";